aggiunta metodi

+++ Slide_370/Voto.cs
@@ -4,6 +4,6 @@
     {
         public DateTime Data { get; set; }
         public Materia Materia { get; set; }
-        public int Valore { get; set; }
+        public double Valore { get; set; }
     }
 }
